feat(integrations): remove tableau as an auth mode replace with two_step
