+++ docs-v2/integrations/all/tableau.mdx
@@ -1,6 +1,6 @@
 ---
-title: Tableau
-sidebarTitle: Tableau
+title: Tableau (PAT)
+sidebarTitle: Tableau (PAT)
 ---
 
 import Overview from "/snippets/overview.mdx"
@@ -14,11 +14,12 @@ import PreBuiltUseCases from "/snippets/generated/tableau/PreBuiltUseCases.mdx"
 ## Access requirements
 | Pre-Requisites | Status | Comment|
 | - | - | - |
-| Paid dev account | ❓ |  |
-| Paid test account | ❓ |  |
-| Partnership | ❓ | |
-| App review | ❓ |  |
-| Security audit | ❓ | |
+| Paid dev account | ✅ Not required | Free, self-signup [here](https://www.tableau.com/products/trial). |
+| Paid test account | ✅ Not required | Create free test account [here](https://www.tableau.com/products/trial). | 
+| Partnership | ✅ Not required | |
+| App review | ✅ Not required | |
+| Security audit | ✅ Not required | |
+
 
 
 ## Setup guide
@@ -32,15 +33,24 @@ _No setup guide yet._
 
 ## Useful links
 
--   [How to generate your Personal Access Token](https://help.tableau.com/current/pro/desktop/en-us/useracct.htm#create-and-revoke-personal-access-tokens)
--   [Auth related docs](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_auth.htm#make-a-sign-in-request-with-a-personal-access-token)
--   [API endpoints](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api.htm)
+| Topic | Links | 
+| - | - | 
+| General | [Website](https://www.tableau.com/) |
+| | [Create a Tableau Cloud trial account](https://www.tableau.com/products/trial) |
+| Developer | [API documentation](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_ref.htm) |
+| | [Authentication guide](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_auth.htm) |
+| | [API version reference](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_versions.htm) |
+| | [Personal Access Token setup](https://help.tableau.com/current/pro/desktop/en-us/useracct.htm#create-and-revoke-personal-access-tokens) |
+| | [Authentication using Personal Access Token](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_auth.htm#make-a-sign-in-request-with-a-personal-access-token) |
 
 <Note>Contribute useful links by [editing this page](https://github.com/nangohq/nango/tree/master/docs-v2/integrations/all/tableau.mdx)</Note>
 
 ## API gotchas
 
-- When creating a new Nango connection, please supply your Token name as your PAT NAME, Token secret as your PAT SECRET, and the content URL, which is the value that appears after `/site/` in the browser address bar for Tableau Cloud, i.e `/site/nango1ec257d4d0/user/`, would be `nango1ec257d4d0`. This field is left blank if you are using Tableau Server.
-- You will also need to provide the [version](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_versions.htm) and [server](https://help.tableau.com/current/api/rest_api/en-us/REST/rest_api_concepts_auth.htm#the-sign-in-uri) name.
+_No reported gotchas yet_
 
 <Note>Contribute API gotchas by [editing this page](https://github.com/nangohq/nango/tree/master/docs-v2/integrations/all/tableau.mdx)</Note>
+
+<Card title="Connect to Tableau (PAT)" icon="link" href="/integrations/all/tableau/connect" horizontal>
+  Guide to connect to Tableau (PAT) using Nango Connect.
+</Card>
