+++ packages/types/lib/providers/provider.ts
@@ -157,10 +157,6 @@ export interface ProviderAppleAppStore extends BaseProvider {
     token_url: string;
 }
 
-export interface ProviderTableau extends BaseProvider {
-    auth_mode: 'TABLEAU';
-}
-
 export interface ProviderBill extends BaseProvider {
     auth_mode: 'BILL';
 }
@@ -213,7 +209,6 @@ export type Provider =
     | ProviderTwoStep
     | ProviderSignature
     | ProviderApiKey
-    | ProviderTableau
     | ProviderBill
     | ProviderGithubApp
     | ProviderAppleAppStore
