+++ packages/webapp/src/pages/Connection/CreateLegacy.tsx
@@ -55,9 +55,6 @@ export const ConnectionCreateLegacy: React.FC = () => {
     const [oAuthClientId, setOAuthClientId] = useState('');
     const [tokenId, setTokenId] = useState('');
     const [tokenSecret, setTokenSecret] = useState('');
-    const [patName, setpatName] = useState('');
-    const [patSecret, setpatSecret] = useState('');
-    const [contentUrl, setContentUrl] = useState('');
     const [organizationId, setOrganizationId] = useState('');
     const [devKey, setDevKey] = useState('');
     const [oAuthClientSecret, setOAuthClientSecret] = useState('');
@@ -201,14 +198,6 @@ export const ConnectionCreateLegacy: React.FC = () => {
                     oauth_scopes_override: oauthSelectedScopes.join(',')
                 };
             }
-        }
-
-        if (authMode === 'TABLEAU') {
-            credentials = {
-                pat_name: patName,
-                pat_secret: patSecret,
-                content_url: contentUrl
-            };
         }
 
         if (authMode === 'JWT') {
@@ -469,19 +458,6 @@ export const ConnectionCreateLegacy: React.FC = () => {
             }
         }
 
-        let tableauCredentialsString = '';
-        if (integration.meta.authMode === 'TABLEAU') {
-            if (patName && patSecret && contentUrl) {
-                tableauCredentialsString = `
-    credentials: {
-        pat_name: '${patName}',
-        pat_secret: '${patSecret}',
-        content_url: '${contentUrl}'
-    }
-  `;
-            }
-        }
-
         let jwtCredentialsString = '';
 
         if (integration.meta.authMode === 'JWT') {
@@ -574,7 +550,6 @@ export const ConnectionCreateLegacy: React.FC = () => {
             !appStoreAuthString &&
             !oauthCredentialsString &&
             !oauth2ClientCredentialsString &&
-            !tableauCredentialsString &&
             !jwtCredentialsString &&
             !tbaCredentialsString &&
             !billCredentialsString &&
@@ -590,7 +565,6 @@ export const ConnectionCreateLegacy: React.FC = () => {
                       appStoreAuthString,
                       oauthCredentialsString,
                       oauth2ClientCredentialsString,
-                      tableauCredentialsString,
                       jwtCredentialsString,
                       tbaCredentialsString,
                       billCredentialsString,
@@ -835,52 +809,6 @@ nango.${integration.meta.authMode === 'NONE' ? 'create' : 'auth'}('${integration
                                             placeholder="Token secret"
                                             optionalValue={tokenSecret}
                                             setOptionalValue={setTokenSecret}
-                                        />
-                                    </div>
-                                </div>
-                            )}
-
-                            {integration?.meta.authMode === 'TABLEAU' && (
-                                <div>
-                                    <div className="flex mt-6">
-                                        <label htmlFor="pat_name" className="text-text-light-gray block text-sm font-semibold">
-                                            PAT Name
-                                        </label>
-                                    </div>
-                                    <div className="mt-1">
-                                        <SecretInput
-                                            copy={true}
-                                            id="pat_name"
-                                            name="pat_name"
-                                            placeholder="PAT Name"
-                                            optionalValue={patName}
-                                            setOptionalValue={setpatName}
-                                        />
-                                    </div>
-                                    <div className="mt-4">
-                                        <label htmlFor="pat_secret" className="text-text-light-gray block text-sm font-semibold">
-                                            PAT Secret
-                                        </label>
-                                        <SecretInput
-                                            copy={true}
-                                            id="pat_secret"
-                                            name="pat_secret"
-                                            placeholder="PAT Secret"
-                                            optionalValue={patSecret}
-                                            setOptionalValue={setpatSecret}
-                                        />
-                                    </div>
-                                    <div className="mt-4">
-                                        <label htmlFor="content_url" className="text-text-light-gray block text-sm font-semibold">
-                                            Content Url
-                                        </label>
-                                        <SecretInput
-                                            copy={true}
-                                            id="content_url"
-                                            name="content_url"
-                                            placeholder="Content Url"
-                                            value={contentUrl}
-                                            setOptionalValue={setContentUrl}
                                         />
                                     </div>
                                 </div>
